backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
anthropic==0.18.0
requests==2.31.0
numpy<2.0
geopandas==0.14.0
shapely==2.0.2
pyproj==3.6.1
pytest==7.4.4
pytest-asyncio==0.23.0
python-dotenv==1.0.0
aiohttp==3.9.1